Fix up various signed/unsigned integer warnings

+++ module/evdi_painter.c
@@ -624,7 +624,7 @@ evdi_painter_connect(struct evdi_device *evdi,
 {
 	struct evdi_painter *painter = evdi->painter;
 	struct edid *new_edid = NULL;
-	int expected_edid_size = 0;
+	unsigned expected_edid_size = 0;
 
 	EVDI_DEBUG("(dev=%d) Process is trying to connect\n",
 		   evdi->dev_index);
@@ -836,8 +836,8 @@ int evdi_painter_grabpix_ioctl(struct drm_device *drm_dev, void *data,
 		}
 	}
 
-	if (cmd->buf_width != efb->base.width ||
-		cmd->buf_height != efb->base.height) {
+	if ((unsigned)cmd->buf_width != efb->base.width ||
+		(unsigned)cmd->buf_height != efb->base.height) {
 		EVDI_ERROR("Invalid buffer dimension\n");
 		err = -EINVAL;
 		goto err_fb;
